backend/requirements.txt
Django==5.0.2
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
cloudinary==1.44.0 